requirements.txt
allure-pytest==2.13.5
pytest==8.3.4
pytest-xdist==3.6.1
python-dotenv==1.0.1
selenium==4.28.1
